pyproject.toml
[tool.poetry]
name = "badk-server"
version = "0.1.0"
description = ""
authors = ["DaProWebStudio"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
djangorestframework = "^3.14.0"
django-ckeditor = "^6.7.0"
django-cleanup = "^8.0.0"
django-debug-toolbar = "^4.2.0"
django-modeltranslation = "^0.18.11"
django-filter = "^23.5"
django-imagekit = "^5.0.0"
django-mptt = "^0.15.0"
transliterate = "^1.10.2"
environs = "^9.5.0"
pillow = "^10.1.0"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
celery = "^5.3.6"
celery-singleton = "^0.3.1"
drf-yasg = "^1.21.7"
django-cachalot = "^2.6.1"
django-redis = "^5.4.0"
python-json-logger = "^2.0.7"
coverage = "^7.3.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
